# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_tmalldedede_agentbox_internal_webhook.CreateWebhookRequest:
    properties:
      events:
        description: 为空则订阅所有事件
        items:
          type: string
        type: array
      secret:
        type: string
      url:
        type: string
    required:
    - url
    type: object
  github_com_tmalldedede_agentbox_internal_webhook.UpdateWebhookRequest:
    properties:
      events:
        items:
          type: string
        type: array
      is_active:
        type: boolean
      secret:
        type: string
      url:
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_webhook.Webhook:
    properties:
      created_at:
        type: string
      events:
        description: 订阅的事件类型
        items:
          type: string
        type: array
      id:
        type: string
      is_active:
        type: boolean
      secret:
        description: 用于签名验证
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  internal_api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:18080
info:
  contact:
    email: support@agentbox.dev
    name: AgentBox Team
    url: https://github.com/user/agentbox
  description: |-
    AgentBox - Open-source AI Agent containerized runtime platform

    ## Overview
    AgentBox provides a unified API for managing AI coding agents (Claude Code, Codex, OpenCode) in containerized environments.

    ## Authentication
    Currently no authentication required. API keys are managed via Credentials API.

    ## API Structure
    - **Public API** (`/api/v1/*`): Core functionality for external integrations
    - **Admin API** (`/api/v1/admin/*`): Platform management operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AgentBox API
  version: "1.0"
paths:
...
  /webhooks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_webhook.Webhook'
            type: array
      summary: 列出所有 Webhook
      tags:
      - webhooks
    post:
      consumes:
      - application/json
      parameters:
      - description: Webhook 配置
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_webhook.CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_webhook.Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: 创建 Webhook
      tags:
      - webhooks
  /webhooks/{id}:
    delete:
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: 删除 Webhook
      tags:
      - webhooks
    get:
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_webhook.Webhook'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: 获取 Webhook 详情
      tags:
      - webhooks
    put:
      consumes:
      - application/json
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook 配置
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_webhook.UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_webhook.Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: 更新 Webhook
      tags:
      - webhooks
swagger: "2.0"
...